Move `failingRefactorTests` to `secure.spec.ts`
